move multiline stat label tweak
