+++ src/main/resources/qolhunters.mixins.json
@@ -11,7 +11,7 @@
     "abilities.MixinInputEvents",
     "ascension.MixinAscensionCrystalObjective",
     "ascension.MixinSoulFlameItem",
-    "betterdescriptions.MixinStatLabelElementBuilder",
+    "multilinestatlabels.MixinStatLabelElementBuilder",
     "bettersoulvalue.MixinClientEvents",
     "bingo.MixinBingoRenderer",
     "bingo.MixinBingoRendererLeaf",
